add version 2.2

+++ src/TcpSocket.h
@@ -262,7 +262,7 @@ public:
 protected:
 	TcpSocket(const TcpSocket& );
 	void OnRead();
-	void OnRead( char *buf, int n );
+	void OnRead( char *buf, size_t n );
 	void OnWrite();
 #ifdef HAVE_OPENSSL
 	/** SSL; Initialize ssl context for a client socket.
